+++ webapp/src/errors.ts
@@ -15,3 +15,21 @@ export class LyraConfigReadingError extends Error {
     super(`Error reading file: [${filename}]`);
   }
 }
+
+export class ServerConfigReadingError extends Error {
+  constructor(filename: string) {
+    super(`Error reading file: [${filename}]`);
+  }
+}
+
+export class ProjectPathNotFoundError extends Error {
+  constructor(projectPath: string) {
+    super(`Project with path [${projectPath}] not found`);
+  }
+}
+
+export class ProjectNameNotFoundError extends Error {
+  constructor(projectName: string) {
+    super(`Project name: [${projectName}] not found`);
+  }
+}