+++ webapp/src/utils/lyraConfig.ts
@@ -1,8 +1,8 @@
 import fs from 'fs/promises';
-import { LyraConfigReadingError } from '@/errors';
 import { parse } from 'yaml';
 import path from 'path';
 import { z } from 'zod';
+import { LyraConfigReadingError, ProjectPathNotFoundError } from '@/errors';
 
 export enum MessageKind {
   TS = 'ts',
@@ -14,7 +14,7 @@ const KIND_BY_FORMAT_VALUE: Record<'ts' | 'yaml', MessageKind> = {
   yaml: MessageKind.YAML,
 };
 
-const configSchema = z.object({
+const lyraConfigSchema = z.object({
   baseBranch: z.optional(z.string()),
   projects: z.array(
     z.object({
@@ -30,23 +30,35 @@ const configSchema = z.object({
   ),
 });
 
-export default class LyraConfig {
+export class LyraConfig {
   private constructor(
     public readonly projects: LyraProjectConfig[],
     public readonly baseBranch: string, // following GitHub terminology target branch called base branch
   ) {}
 
+  public getProjectConfigByPath(projectPath: string): LyraProjectConfig {
+    const projectConfig = this.projects.find(
+      (project) => project.path === projectPath,
+    );
+    if (projectConfig) {
+      return projectConfig;
+    }
+    throw new ProjectPathNotFoundError(projectPath);
+  }
+
   static async readFromDir(repoPath: string): Promise<LyraConfig> {
+    // TODO: cache this call with TTL
     const filename = path.join(repoPath, 'lyra.yml');
     try {
       const ymlBuf = await fs.readFile(filename);
       const configData = parse(ymlBuf.toString());
 
-      const parsed = configSchema.parse(configData);
+      const parsed = lyraConfigSchema.parse(configData);
 
       return new LyraConfig(
         parsed.projects.map((project) => {
           return new LyraProjectConfig(
+            project.path,
             KIND_BY_FORMAT_VALUE[project.messages.format],
             path.join(repoPath, project.path, project.messages.path),
             path.join(repoPath, project.path, project.translations.path),
@@ -60,8 +72,9 @@ export default class LyraConfig {
   }
 }
 
-class LyraProjectConfig {
+export class LyraProjectConfig {
   constructor(
+    public readonly path: string,
     public readonly messageKind: string,
     public readonly messagesPath: string,
     public readonly translationsPath: string,